\name{fancypackage-package}
\alias{fancypackage-package}
\alias{fancypackage}
\docType{package}
\title{\packageTitle{fancypackage}}
\description{\packageDescription{fancypackage}}
\section{Package Content}{\packageIndices{fancypackage}}
\author{\packageAuthor{fancypackage}}
\section{Maintainer}{\packageMaintainer{fancypackage}}
\keyword{package}
